Removed STATIC_ROOT entry in .env
This is None by default so it should be left out of the .env.
The Dockerfile sets it to /static as of now and the docker-compose
file sets /static as a volume between the Nginx and Django
container. This would mean that Prod sets it to the appropriate
values when needed. Therefore it is not required in .env.