Send pushover notification on change detection (#4)
* Added .env for Pushover.

* Pushover vars commented out by default in .env

By default these should be None so that the system can detect when
they are not set. I commented them out in the .env.example so if
someone just copies it over, by default it is not enabled.

* Removed STATIC_ROOT entry in .env

This is None by default so it should be left out of the .env.
The Dockerfile sets it to /static as of now and the docker-compose
file sets /static as a volume between the Nginx and Django
container. This would mean that Prod sets it to the appropriate
values when needed. Therefore it is not required in .env.

* Created send_pushover function

This will send a notification on Pushover. Added a single basic
test to ensure that a simple notification works. Maybe in the
future we should check to see that attachments are sent
successfully.

* Send Pushover notification on change detection

For now it's just pushover. In the future, we should look at
making this something that can be modified on a settings page or
something.

* Uncomment .env.example Pushover Settings

Looks like GitHub Actions depend on the .env.example and we need
to uncomment the Pushover settings to ensure that the tests pass.

* Uncomment .env.example Pushover Settings

Looks like GitHub Actions depend on the .env.example and we need
to uncomment the Pushover settings to ensure that the tests pass.